chore(deps): Update dev dependencies with `npm audit fix`
This updates several non-load-bearing "dev dependencies" that we use in our
testing framework to resolve security warnings that appear when running `npm
audit`.  Due to the nature of the packages' use inside of this reposoitory,
none of these updates respresents a risk to users of the router: they are
not included in its compiled distribution nor used at runtime.