Point developers to `.with_metrics()` in metrics assertions
It's easy to forget `.with_metrics()` in an async metrics test, and you
have to read dev-docs/metrics.md to know about this.

So let's add the link to dev-docs/metrics.md in the doc comments for the
metrics assertion macros, and also call it out in the panic message that
you get if you forget to do this.