update CircleCI base image
This commit fixes an issue with a downstream dependency
(`aws-lc-sys`) caused by using an old version of GCC by updating
the base CircleCI image to `current`.

This was realistically a long-time coming because the previous
tag of `stable` was deprecated in favor of `current` back in 2022.

Below is a quote from the [docker hub page](https://hub.docker.com/r/cimg/base)
explaining the change to the tag names:

| current - This image tag points to the latest, production ready
| base image. This is a replacement for the old stable tag. This
| image should be used by projects that want a decent level of
| stability but would like to get occasional software updates.
| It is typically updated once a month.