Fix remaining broken tests
I've fixed the mock expectations (which had changed in the
implementation) in the query_planner tests. The one expecting the
closed_service seems like a weird test, so not really sure what to do
about that one. I think the back-pressure improvements may mean we don't
even need that test...

I've re-enabled the co-processor test since our ongoing implementation
work means that test no longer fails.

In general, all tests are passing on my laptop, so this is now close to
review-able.