Limit the depth of introspection queries
The [schema-intropsection schema](https://spec.graphql.org/draft/#sec-Schema-Introspection.Schema-Introspection-Schema)
is recursive: a client can query the fields of the types of some other fields,
and so on arbitrarily deep.
This can produce responses that grow much faster than the size of the request.

To protect against abusive Router now refuses to execute introspection queries
that nest list fields too deep and returns an error instead.
The criteria matches `MaxIntrospectionDepthRule` in graphql-js,
but may change in future versions.

In case it rejects legitimate queries,
this check can be disabled in Router configuration:

```yaml
supergraph:
  introspection: true  # Needed to enable schema introspection in the first place
  introspection_check_max_depth: false  # Defaults to true
```