wrap up `mock_subgraph_service_withf_panics_should_be_reported_as_service_closed` test
This appears to test an intended-enough behaviour.

Originally introduced: https://github.com/apollographql/router/pull/918
Comment outdated as of: https://github.com/apollographql/router/pull/1440

This is a behaviour change compared to #1440, but the intended behaviour
from #918.

What's a bit worrying is that this depends on the internal composition of
our pipeline, but is externally observable?