with_metrics() doesn't require Send
The Router service is required to use UnsyncBoxBody because
futures::Stream is not Sync. Changing this to use a different Streams
implementation (e.g.: tokio) is a lot of work and not guaranteed to
improve things.

This causes issues for `with_metrics()` since it requires the provided
closure to be `Send` so that it can be boxed(). Actually,
`with_metrics()` doesn't have to box the created closure, it could use
`boxed_local()`.

That leads to a simple change to `with_metrics()` which means Router
service can be used with it and all existing tests continue to function.