Update various existing health tests and migrate some to plugin
Two of the health check tests now make more sense when tested against
the new plugin, so I've migrated those.

The remaining three tests stay in axum_factory/tests.rs, although I had
to modify one to manually insert endpoint, since this is now only done
by the plugin which isn't present in the test.

It might be better to just delete that test, but leaving for now.