Tweak the defaults a little bit to try and be more k8s friendly
By default, k8s has the following defaults:

```
failureThreshold:3
periodSeconds: 10
```

So, by default, it would take between 212 and 30 seconds to recognise a pod is not ready. So, by default, we should set our sample interval to 10 second and our marked unready to 30 seconds (or 3x sample interval) and that would make it useful out of the box.

Tweaking via configuration allows non-default values to be assigned.