Remove the ability to auto-upgrade a router configuration at runtime
In Router 1.x, during configuration parsing/validation the router
upgrade migrations could be applied "auto-magically" to generate a valid
runtime representation of a config for the life of the executing
process.

This can be problematic, since it makes de-bugging and support more
challenging than we'd like.

We want to retain the ability to upgrade a configuration using
migrations, but we don't want to auto-magically do this during runtime.
It should be the user's decision to make an explicit upgrade.

This is what this PR does. Various impacted tests are updated.

I also updated the link to the "migration" guide to be a link to the
"upgrade" guide at its new location.