change error message for unsupported options for `router config upgrade`
It doesn't make sense to point users to `router config upgrade` when this is only run during `router config upgrade`. Since we are starting from a clean slate in router 2.0.0, the effective_migrations are also all unsupported, rather than deprecated, options, and should be emitted as an error than a warn.