Prevent deadlocks on metrics exporter shutdown
The aggregate meter provider had not taken into account the possibility that we would add metrics directly to metrics exporters. This meant that if a meter has metrics in them then the circularity would cause a deadlock.

This only happens on shutdown, but I have added tests for reload just in case.