Remove incomplete `Pending`-based attempt at compute backpressure
Since version 1.58.0, Apollo Router has a dedicated thread pool
with a priority queue for computation-heavy jobs:
parsing + validation, query planning, and introspection.

Initially we tried to exert back-pressure by returning `Poll::Pending`
in `tower::Service::poll_ready` for the query planner service
when the queue is full above a certain threshold.
But `Pending` is supposed to represent waiting _for something_,
and we’re supposed to register a "waker" on the async context
to notify Tokio to poll this task again when that thing happens.
Additionally we’ve found that Tower-based back-pressure
is completely broken in 1.x, and made significant changes in Router 2.x to fix it.

Backpressure like in 2.x is not gonna happen in 1.x,
so remove the incorrect usage of `Pending` instead.
This makes the compute job queue effectively unbounded.