chore: Delegate setting "latest" to RELEASE_CHECKLIST.md, rather than CI.
Because we now have two different "major" versions on the Router
repository, we need to make sure they don't both compete for which is
"latest".  Removing this from CI and moving this to an explicit step in
`RELEASE_CHECKLIST.mdd` is a temporary solution and we can figure out how to
operationalize it further next. (Which could actually involve re-introducing
this commit, but only on the 2.x line, or something, but there are also
other signals we could automatically react to.)