add dhat heap profiling for Supergaph and QueryPlanner functions
Building on the work in https://github.com/apollographql/router/pull/6743, this commit adds heap profiling tests for a few of the apollo_federation::Supergraph and apollo_federaion::query_plan::query_planner functions:
* apollo_federation::Supergraph::new()
* supergraph.to_api_schema()
* supergraph.extract_subgraphs_to_supergraph()
* apollo_federation::query_plan::query_planner::QueryPlanner::new()
* planner.build_query_plan()

This is not meant to be a comprehensive heap profiling testing framework, but merely an indication to us that something has dramatically changed in the most used query planning functionality. A smoke signal, if you will.