add dhat heap profiling tests for Supergaph and QueryPlanner functions (#6814)
Building on the work in https://github.com/apollographql/router/pull/6743, this commit adds heap profiling tests for a few of the apollo_federation::Supergraph and apollo_federaion::query_plan::query_planner functions:
* apollo_federation::Supergraph::new()
* supergraph.to_api_schema()
* supergraph.extract_subgraphs_to_supergraph()
* apollo_federation::query_plan::query_planner::QueryPlanner::new()
* planner.build_query_plan()

This is not meant to be a comprehensive heap profiling testing framework, but merely an indication to us that something has dramatically changed in the most used query planning functionality. A smoke signal, if you will.

Heap usage testing with the `dhat` crate is a bit finicky, and you can't really add them to an existing test file or combine multiple tests in a file ([more on this in the crate docs](https://docs.rs/dhat/latest/dhat/#heap-usage-testing)). Tests have to be added as individual tests, or we have to create a custom test harness. We already have a whole bunch of different test harnesses in this repo, which make it quite confusing to pick _which_ harness/integration test "framework" to write additional integration tests with. None of the existing ones really work for this kind of testing. _So_, in order to make these tests as easy as possible to reason about, they are added as their own individual integration tests in `apollo-federation`'s `Cargo.toml`. You will see them like as regular integration tests when you run `cargo nextest run`:

<img width="887" alt="Screenshot 2025-02-17 at 17 57 41" src="https://github.com/user-attachments/assets/b91b372c-66b4-4db0-a6bd-7d27d15f9ac6" />

For the time being, the tests use a very tiny, not very complex supergraph and operation. It will be good to add additional tests with a more complex version of both, as the heap usage is going to be vastly different and will scale differently.