Ensure `build_query_plan()` cancels when the request cancels (#6840)
When a request is cancelled, resources (compute time/memory) allocated to that request should be released. However, query planning is an exception, and currently keeps running after the request is cancelled. This PR adds cooperative cancellation to `build_query_plan()`, to ensure that when the request is cancelled, query planning errors shortly after.

Co-authored-by: Sachin D. Shinde <sachin@apollographql.com>