Pattern for iterating over directive applications (#6943)
Implements a pattern for iterating over all applications of a particular directive. In JS, this is called as directive.applications() on some directive definition. In Rust, the directive definition itself doesn't have a pointer back to the FederationSchema, so the schema will return the iterators. There is a different iterator for each directive because the contained fields don't have useful common types, and this tries to provide a simply, strongly-typed directive representation for consumption.