Replace Rhai specific hot-reload functionality with general hot-reload
The Rhai hot-reload functionality in 1.x only worked because 1.x created
the entire service stack for each request.

In 2.0, we no longer do that and so the Rhai hot-reload functionlity
does not work.

The solution is to remove the rhai specific hot-reload code and to add
rhai source code into the set of things watched by the router which will
trigger a hot reload event.