Add exemption for RUSTSEC-2024-0437 and prevent adding new vulnerable dependencies
[RUSTSEC-2024-0437](https://rustsec.org/advisories/RUSTSEC-2024-0437)
shows a vulnerability in the `protobuf` crate's decoder implementation.

We use protobuf 2.x through the `prometheus` and `opentelemetry-prometheus` crates.
A patch is only available for protobuf 3.x.
We basically cannot update to a patched version of `protobuf`,
especially since we are also using an older version of opentelemetry.

Luckily, the prometheus crates only use the protobuf *encoder* implementation,
not the decoder implementation. So we are not actually affected by the
advisory.

This patch adds an exemption for the vulnerability, and disallows adding
any *new* dependency on the vulnerable protobuf crates, so we can not
silently add a dependency that *does* use the decoder implementation.