Fix flaky test for timed-out operation name in tracing
This test is frequently failing on CI, especially on macOS.

Looking into it, the test is making an incorrect assertion. The purpose
of the test is to make sure that when a request times out, it has the
GraphQL operation name in the `otel.name` attribute.

This assertion was changed in the backpressure work, likely because
backpressure made the test fail:

https://github.com/apollographql/router/commit/8f1459b59eb77dade3fa559649735b0ca7aced12#diff-922f40380fab15ec672ef4074c95504d1e925de392bb8f11b38f83a080d83a58L116-R117

Note this *does not* indicate broken behaviour in the router: the test
just failed because it specifies an extremely short timeout of 1
nanosecond. When backpressure functions correctly, the request times
out before the GraphQL operation is even parsed. So we don't know what
the operation name is.

To fix the test, we have to make sure that the GraphQL operation is
parsed *before* we time out. So I increase the timeout values here.

As with any time-based test, this could still be flaky, in that case we
have to increase the timeouts a bit more.