Remove the cancellation changeset
Per #6949, this feature doesn't yet work fully. We also need
significantly more design and testing time in order to release it
without breaking customers.

The part of the feature that's already implemented doesn't hurt. It's
not a problem to release it, it just won't do much.

But we shouldn't advertise it in the changelog for 2.1.