Add a timeout on waiting for a connection to close after a graceful shutdown.
We are sometime seeing connections being hung onto after graceful shutdown. Add a 60 second timeout, particularly with http2 connections, but this has been seen with http1.

This PR adds a hard coded 60 second timeout after a connection has started shutting down for it to be terminated. If it doesn't close properly then it will be dropped.

If we find that this fixes the connection retention issues then we will make this configurable.