Adjust compute job worker pool queue size
The compute job worker pool is used for CPU-bound tasks, like GraphQL
parsing, validation, and query planning. When there are too many jobs to
handle in parallel, jobs enter a queue.

We previously set this queue size to 20 (per thread) somewhat arbitrarily.
We got some signals that this may be too little. Actually, most jobs complete
very quickly.

This patch increases the queue size to 1 000 jobs per thread. For
reference, in older router versions that were based on the JavaScript
query planner, the equivalent queue size was _10 000_.

The number is still a bit arbitrary. Along with some other tweaks to job
priorities we expect this to give better behaviour and reject fewer
requests needlessly.

<!-- [ROUTER-1236] -->