Fix indirect options bug where we mistakenly avoid caching for `@skip`/`@include` elements (#7239)
We had a bug in the indirect options computation relative to the JS code, where the JS code was not caching the computation if the subsequent operation element had `@skip`/`@include`, but the Rust code forgot to return and effectively was caching (and was doing the computation twice). After looking at the code more closely, both the JS and Rust code are wrong, and there’s nothing wrong with caching indirect options when the subsequent operation element has `@skip`/`@include` (because the computation itself doesn't depend on the subsequent operation element at all). This PR accordingly just removes the branch.