Avoid unnecessary cloning in the deduplication plugin
We now only duplicate if there is a waiter. I think this also fixes an
undiagnosed race condition where an entry could remain in the wait map
AFTER the notifications have been sent.

This is now fixed by explicitly waiting for the spawned task to complete
before checking for receivers and sending the notification to them.