Make `SelectionTrie::extend` cheaper when fewer collisions.
Using an `Arc` reference for the values of the
`SelectionTrie::selections: IndexMap<String, Arc<SelectionTrie>>` means
we can cheaply extend or merge `SelectionTrie` instances that do not
have high overlap, because non-overlapping keys (and the entire subtries
beneath them) can be cheaply cloned across tries.

This cloning is copy-on-write (thanks to `Ref::make_mut`), so modifying
the new trie triggers lazy clones of borrowed `Arc<SelectionTrie>`
references, but does not affect the original trie.