Add FederationError variant for subgraph errors (#7436)
There are some places in composition code that need to create a `FederationError`, but annotated with subgraph name. Ideally we would split federation errors into internal errors, query planning errors, and composition errors, and give composition errors the ability to be annotated with subgraph name. But that's further out, so for now, this PR tries to introduce a small changeset for allowing subgraph name annotation.