chore: tweak mise usage on CI
Remove usage of `mise trust`, which is not required on CI.

Add the `--raw` flag to `install`, to hopefully handle random hangs.
This flag disables parallel installs, but importantly disables
the terminal UI with progress bars etc. that mise uses. Without
the progress bar, CircleCI should be able to detect that there
has been no output for a while and kill the job if it gets stuck.