chore: simplify yaml config in traffic shaping tests
I was reading this for unrelated reasons and was a bit horrified that
we're using string concatenation to merge config in a language with
significant whitespace. I think there's not much benefit to factoring
out this config, and it makes the tests harder to read, so I've just
copy-pasted it into each usage site.