Temporarily disable cargo-fuzz
There seem to be some sporadic `mise install` failures happening in CI that
seem directly attriubted to `cargo-fuzz`.  There are a few theories here,
but to just eliminate them for now, we'll try disabling it and not
installing it.  This should give us an angle on which to figure out next
steps.

Let's give it a week.