chore: update tracing-serde
Because it's part of the `tracing-` dependency group, renovate won't
update this independently.

`tracing-subscriber` already uses `tracing-serde` 0.2.0, so upgrading
our own dependency removes a duplicate crate from the dep tree.

The breaking change is in a part of the API that we don't use, and it
does not affect the serialization format: https://github.com/tokio-rs/tracing/releases/tag/tracing-serde-0.2.0