disabling the health endpoint in 2.0 should be supported (#7519)
In 2.x the health check endpoint support was converted to be a new plugin. Unfortunately, the support for disabling the endpoint was lost during the conversion.

This is now fixed and a new unit test ensures that disabling the health check does not result in the creation of a health check endpoint.