deps: pin hyper-util
https://github.com/apollographql/router/pull/7159 doesn't quite work and
I'm not sure how to solve it right now. It might be a hyper-util bug but
it might also be on our end. It compiles and runs fine but the tracing
tests fail.

Our `test_updated` job is also failing because it's attempting to use
hyper-util 0.1.11. By pinning the dependency that should start working
again.