Remove `tracing::event!` calls inside tracing layres
Using `tracing::error!()` or other event macros inside a
tracing-subscriber layer callback may just drop the event.

Elsewhere, we've solved this by manually building an event and using
`context.event()`.
https://github.com/apollographql/router/blob/678dee7e1a85d4520d84e90ef0913b892a55ee65/apollo-router/src/plugins/telemetry/reload.rs#L175-L187

I think this is not appropriate to do in this case. If an error
occurred inside the tracing-subscriber layer, we shouldn't expect to be
able to record new events with `tracing`. So I simply changed these
calls to `eprintln!()`. The messages will at least go _somewhere_.

I think there's an argument that these could be panics, but I don't
want to risk breaking the world.

<!-- [ROUTER-1016] -->