Take advantage of improvement in the http crate for cloning
In an older version of the crate, `http::Extensions` were not `Clone`. This
had the consequence that `http::Request` and `http::Response` were also
not clone. To work-around this limitation, we had functions in our
http_ext module which would "clone" a `Request` or a `Response`, but
they'd omit `Extensions`.

Generally speaking, this wasn't a problem for most use cases. However,
over time we started to add functionality, such as file uploads, to the
router which relied on the correct behaviour of cloning and expected
Extensions to be preserved across a clone.

This finally resulted in an issue for a customer who could not
understand why modifying a request in a Rhai script would cause a File
Upload to fail.

It transpires that at version 1.0.0 of the http crate, an incompatible
change was made which forces `http::Extensions` to be `Clone` which means
both `http::Request` and `http::Response` are now `Clone`.

I've replaced our `http::Extensions` ignoring implementation with the
provided `Clone` behaviour and this means that Request and Response
Extensions are now correctly cloned. I've added tests for this as well.