chore(ci): codecov on macOS in CI nightlies
This makes CI work on macOS on *nightly* builds.  Overall, the process of
running codecov on Rust codebases of this size is pretty slow.

It roughly doubled the test run on macOS when it was part of the tests
directly, and it wouldn't even run on most of our Linux boxes due to the
amount of memory that llvm-cov chews up in the process.

To account for that, but still get some coverage numbers, we'll go ahead and
introduce this as a nightly step that lets us get the code coverage once a
day, or on demand as we need it.  To do it on demand on a particular pull
request, this commit also introduces a "triggerable" job on CircleCI called
"coverage_only".

Developers can of course run codecov locally using `cargo llvm-cov nextest`.