Revert "xtask release process (#5275)"
This reverts commit 72d8d9b227b2f1c33c37948b3146774c33738a65.

While a short-term attempt to avoiding copying and pasting some commands, this
process just actually doesn't work, and it's incompatible with moving to a
fully CI-based workflow based on some of its design choices.

As the person making the releases, I can attest that it's more error prone
than what we had right now, and the failure conditions are greater.