Fix the default endpoint for otlp over http metrics in 2.x
We use the rust opentelemetry crate to export metrics and traces.
In 1.x we use v 0.13 of that crate.
In 2.x we use v 0.17 of that crate.

It seems as though in the earlier version, exporting otlp metrics over
http would work without the correct path (/v1/metrics). This does not
work in v0.17. This means that exporting metrics over http has never
worked in router 2.x

I've added code to explicitly force the path to be correct (as we
already do for traces).