Move the `ParsedDocument` hack
We've been looking at refactoring parts of the pipeline for a long time.
This small change should solve a big blocker to making tweaks to the pipeline.

Whenever I try to move something to a more sensible place, all the tests
start failing, because the supergraph service must have a
`ParsedDocument` in context (it's inserted by the router service), but
none of the tests put it in context. Actually, tests usually **cannot**
put it in the context, because to do so, they need to be able to parse
the query against the schema. When using `TestHarness`, you do not have
access to the schema.

Solving that completely will be difficult. But we can make things a bit
better. The hack is currently applied _right before_ query planning.
We can instead apply it inside the `TestHarness`, which more closely
mimicks the real-world behaviour: the ParsedDocument is inserted before
we reach the supergraph service.

Now, we can freely move things around inside the supergraph service
without breaking a load of tests that rely on this hack.

This breaks one test, which was testing that the supergraph service
raised an error on invalid GraphQL. The supergraph service does not
handle invalid GraphQL, so this was testing a nonsensical behaviour. I
moved that test to use the router service.