chore(federation): Use `try_from()` instead of `try_into()` to avoid having to write explicit type signatures (#7609)
I got some feedback in a previous PR that we could avoid manually having to specify type signatures in a `match` with `try_into()` if we just used `try_from()` instead, and realized we had that issue in a few different places. This also avoids a particular IntelliJ issue where it wouldn't be able to infer types in these cases.