chore(xtask): pass through test name filters
Just a small enhancement, if you use `cargo xtask test` you can pass a
filter in command-line arguments, just like with `cargo test` and `cargo
nextest run`.

Ideally we'd just support any and all arguments to `cargo test` and
`cargo nextest run` here. But I'm not sure how to achieve that and this
is the one I miss the most often by far.