Batching is not supported for GET requests
When sending an operation as a GET request, like `/graphql?query={ user { id } }`,
you cannot send a batch request.

We use `serde_urlencoded` to parse such requests, and the behaviour is
like this:
```rust
use serde_json::Value;
use serde_json::json;

assert_eq!(
    serde_urlencoded::from_str::<Value>("query=graphql").unwrap(),
    json!({ "query": "graphql" }),
);
assert_eq!(
    serde_urlencoded::from_str::<Value>("query=graphql&query=other").unwrap(),
    json!({ "query": "other" }),
);
assert_eq!(
    serde_urlencoded::from_str::<Value>("query[]=graphql&query[]=other").unwrap(),
    json!({ "query[]": "other" }),
);
```

That is, there is no way for `batch_from_urlencoded_query` to ever
parse a `serde_json::Value::Array` from the query string.

That means that parsing a batch from a query string is essentially a
no-op, and I suggest we remove the code for it.