Use arc_swap crate to perform an atomic swap on the cache
The cache is updated infrequently, read often. We can use ArcSwap to
perform an atomic update on the cache when change it.

ArcSwap isn't Clone, that would make it fairly useless, so when we clone
our Factory, they get their own instance of ArcSwap with a clone of the
current state of the cache.

My feeling is that the additional memory used from the cache duplication
will be small, but the performance benefit from not locking the cache
may be significant. We'll see...