new error container for connector errors
instead of using the more generic `graphql::Error` type, that we don't
want to bring into apollo-federation, this is a more specific type that
can be converted to `graphql::Error` at the end of mapping.

this is not the ideal API, but it unblocked moving the mapping code into
apollo-federation.