pq: include operation name in "PQ not found" error
We have found that during migrations to PQ-based safelisting, it is
helpful to have a bit more information about what operation the client
is trying to run than just the hashed PQ ID, to help find the client app
that is being blocked. While the client name can be found in the
context, the operation name (if provided explicitly in the body's
`operationName` field by the client) is helpful. We now provide this as
an extension on the error.

(It is both directly helpful in that it is available in the JSON error
response if you're looking at individual failures, and it also makes it
easy for a router_service-level Rust or Rhai plugin that's trying to
provide more observability into PQ errors to look at the error for the
extra data. This error occurs before supergraph_service runs, which
means that the operation name is not in an easily accessible place for
plugins.)