chore: pin serde_yaml to 0.8.x range
Version 0.9.x has breaking changes and the underlying package is
deprecated. The package works as-is and we can consider replacing
it in the future. We'll allow 0.8.x just in case there are any
security updates that we need to be aware of.

See: https://github.com/dtolnay/serde-yaml/releases/tag/0.9.34

This resolves the Renovate complaints about this package on the
Dependency Dashboard: https://github.com/apollographql/router/issues/57