Unify `NamedSelection` enum variants into single struct.
Important note: this refactoring enables the `...` spread operator in
`NamedSelection`, since it seemed important to design that near-future
syntax at the same time as I was redesigning the `NamedSelection`
representation. I believe this new `...` syntax does not need
`ConnectSpec` version gating because it appears only in a `parse_v0_3`
method which already requires/assumes connect/v0.3.