chore: pin `rand` version used for fuzzing
This `rand` version must be aligned with the `bnf` version used, as we
pass a random number generator to `bnf::Grammar::generate`.

In the next version of `bnf`, we can remove our `rand` dependency thanks
to https://github.com/shnewto/bnf/pull/175.

Until then, we should pin it with `=` syntax so renovate stops making
PRs for it that can't be merged.