Enable parsing of `...` spread syntax.
This `...` token is an optional prefix for NamedSelection elements, and
corresponds to the `NamingPrefix::Spread(Option<Range<usize>>)` variant.

Note that we already have precedent for `NamingPrefix::Spread(None)`,
which is triggered when a PathSelection has a trailing SubSelection that
can be inlined into a larger SubSelection (which is behavior we have
already shipped). Functionally, the `...` version of that behavior is
the same (except that it can spread paths that do not have a trailing
SubSelection), so this commit modifies the parser to accept `...` (and
record its source range), with tests of the new syntax, but the
downstream `apply_to_path` and `compute_output_shape` machinery to
handle `NamingPrefix::Spread` is already in place.