Undo possibility of `StringTemplate::from_str` failing.
This failure case was helpful for surfacing indirect uses of
`StringTemplate::from_str` in places where a `ConnectSpec` version was
required to parse JSONSelection expressions within the template.

However, the failure could be a surprising new case for consuming code,
so this commit puts behavior back the way it was before (infallible).