Keep `ConnectSpec::{latest,next}` configured as `#[cfg(test)]`.
These functions can be useful in tests, but they are risky in non-test
code because they encourage depending on something that could change
suddenly, whenever a new ConnectSpec version is released.

This commit also removes the `impl Default for ConnectSpec`, since it
was another easy way to accidentally use `ConnectSpec::latest()` outside
tests.